requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.11.3
alembic=1.15.2=pypi_0
altgraph=0.17.4=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
apscheduler=3.11.0=pypi_0
autocommand=2.2.2=pypi_0
backports-tarfile=1.2.0=pypi_0
bcrypt=3.2.0=py311h827c3e9_2
blinker=1.9.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
certifi=2025.4.26=pypi_0
cffi=1.17.1=py311h827c3e9_1
charset-normalizer=3.4.2=pypi_0
cheroot=10.0.1=pypi_0
cherrypy=18.10.0=pypi_0
click=8.1.8=py311haa95532_0
colorama=0.4.6=py311haa95532_0
contourpy=1.3.2=pypi_0
cryptography=44.0.1=py311hbd6ee87_0
cycler=0.12.1=pypi_0
distro=1.9.0=pypi_0
downloads=1.0.0=pypi_0
et-xmlfile=2.0.0=pypi_0
filelock=3.18.0=pypi_0
flask=3.1.0=pypi_0
flask-cors=5.0.1=pypi_0
flask-jwt-extended=4.4.4=py311haa95532_0
flask-migrate=4.1.0=pypi_0
flask-sqlalchemy=3.1.1=pypi_0
fonttools=4.57.0=pypi_0
fsspec=2025.3.2=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
greenlet=3.2.1=pypi_0
h11=0.16.0=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
iniconfig=2.1.0=pypi_0
itsdangerous=2.2.0=py311haa95532_0
jaraco-collections=5.2.1=pypi_0
jaraco-context=6.0.1=pypi_0
jaraco-functools=4.2.1=pypi_0
jaraco-text=4.0.0=pypi_0
jinja2=3.1.6=py311haa95532_0
jiter=0.9.0=pypi_0
kiwisolver=1.4.8=pypi_0
lazy-loader=0.4=pypi_0
libffi=3.4.4=hd77b12b_1
mako=1.3.10=pypi_0
markupsafe=3.0.2=py311h827c3e9_0
matplotlib=3.10.3=pypi_0
more-itertools=10.7.0=pypi_0
mpmath=1.3.0=pypi_0
networkx=3.4.2=pypi_0
numpy=2.2.5=pypi_0
openai=1.77.0=pypi_0
opencv-python=4.11.0.86=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.0.16=h3f729d1_0
packaging=25.0=pypi_0
pandas=2.2.3=pypi_0
pefile=2023.2.7=pypi_0
pillow=11.2.1=pypi_0
pip=25.1=pyhc872135_1
pipa=0.1.0=pypi_0
piqa=1.3.2=pypi_0
pluggy=1.5.0=pypi_0
portend=3.2.1=pypi_0
psutil=7.0.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.11.4=pypi_0
pydantic-core=2.33.2=pypi_0
pyinstaller=6.13.0=pypi_0
pyinstaller-hooks-contrib=2025.4=pypi_0
pyjwt=2.10.1=py311haa95532_0
pymysql=1.1.1=pypi_0
pyopenssl=25.1.0=pypi_0
pyparsing=3.2.3=pypi_0
pytest=8.3.5=pypi_0
python=3.11.11=h4607a30_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2025.2=pypi_0
pywin32-ctypes=0.2.3=pypi_0
pyyaml=6.0.2=pypi_0
requests=2.32.3=pypi_0
routers=0.10.1=pypi_0
scikit-image=0.25.2=pypi_0
scipy=1.15.3=pypi_0
seaborn=0.13.2=pypi_0
setuptools=78.1.1=py311haa95532_0
six=1.17.0=py311haa95532_0
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.40=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.14.0=pypi_0
tempora=5.8.1=pypi_0
thop=0.1.1-2209072238=pypi_0
tifffile=2025.6.11=pypi_0
tk=8.6.14=h0416ee5_0
torch=2.7.0=pypi_0
torchvision=0.22.0=pypi_0
tqdm=4.67.1=pypi_0
typing-extensions=4.13.2=pypi_0
typing-inspection=0.4.0=pypi_0
tzdata=2025.2=pypi_0
tzlocal=5.3.1=pypi_0
ultralytics=8.3.129=pypi_0
ultralytics-thop=2.0.14=pypi_0
urllib3=2.4.0=pypi_0
utils=1.0.2=pypi_0
vc=14.42=haa95532_5
vs2015_runtime=14.42.34433=hbfb602d_5
werkzeug=3.1.3=py311haa95532_0
wheel=0.45.1=py311haa95532_0
xlsxwriter=3.2.5=pypi_0
xz=5.6.4=h4754444_1
zc-lockfile=3.0.post1=pypi_0
zlib=1.2.13=h8cc25b3_1
